the scrap button doesnt crash when there is no event

+++ app/src/main/java/com/google/eventure/EventActivity.java
@@ -65,15 +65,19 @@ public class EventActivity extends ScheduleActivity {
                 textEvent.setText(null);textNotifHour.setText(null);
                 textNotifMinute.setText(null);
             }
+
+
         buttonScrap.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Toast.makeText(getApplicationContext(), "Event Deleted.", Toast.LENGTH_LONG).show();
 
+            if(ev!=null) {
                 db = new DatabaseHelper(getApplicationContext());
                 db.deleteEvent(ev.getId());
                 Intent EventIntent = new Intent(EventActivity.this, ScheduleActivity.class);
                 startActivity(EventIntent);
+                Toast.makeText(getApplicationContext(), "Event Deleted.", Toast.LENGTH_LONG).show();
+            }
             }
         });
 
@@ -96,28 +100,24 @@ public class EventActivity extends ScheduleActivity {
                     int Repeat = Integer.parseInt((textRepeat.getText().toString()))+1;
 
                     //create a new list of events to be added
-                     ArrayList<Event> toAdd = new ArrayList<>(Repeat);
+                    if(ev==null) {
+                        for (int i = 0; i < Repeat; i++) {
+                            // then create the event
+                            Event1 = new Event(year, month, (day * 7 * i), SHour, SMinute, EHour, EMinute, NHour, NMinute,
+                                    textNote.getText().toString(), textEvent.getText().toString(), textLoc.getText().toString());
 
-                    for (int i = 0; i < Repeat; i++) {
-                    // then create the event
-                    Event1 = new Event(year, month, (day + 7 *i), SHour, SMinute, EHour, EMinute, NHour, NMinute,
-                            textNote.getText().toString(), textEvent.getText().toString(), textLoc.getText().toString());
-                         toAdd.add(Event1);
+                            db = new DatabaseHelper(getApplicationContext());
+                            Event1.setId(db.createEvent(Event1, new long[]{student.getID()}));
+                        }
                     }
-                    db = new DatabaseHelper(getApplicationContext());
-
-                    //   for (Event e : toAdd) {
-                    // add the event to database
-                    if (ev == null)
-                        Event1.setId(db.createEvent(Event1, new long[]{student.getID()}));
-                    else {
-                        ev.setEminute(EMinute);
-                        ev.setName(textEvent.getText().toString());ev.setYear(year);
-                        ev.setlocation(textLoc.getText().toString());ev.setmonth(month);
-                        ev.setday(day);ev.sethour(SHour);ev.setMinute(SMinute);
-                        ev.setEhour(EHour);
-                        ev.setEminute(EMinute);
-                        db.updateEvent(ev);
+                    else{
+                            ev.setEminute(EMinute);
+                            ev.setName(textEvent.getText().toString());ev.setYear(year);
+                            ev.setlocation(textLoc.getText().toString());ev.setmonth(month);
+                            ev.setday(day);ev.sethour(SHour);ev.setMinute(SMinute);
+                            ev.setEhour(EHour);
+                            ev.setEminute(EMinute);
+                            db.updateEvent(ev);
 
                     }
                   //set values back to null
